refactor(bootstrap): keep module internal
